Move signals module into the process module
Also renames Receive to ReceiveSignal, to match ReceiveSignals and to
reflect it's no longer called signals::Receive.

The old signals module is replaced with a reexport of the types from the
process module, using the old types, with a deprecated warning.
